pyproject.toml
[project]
name = "underground-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.100.0",
    "httpx==0.25.0",
    "hypercorn==0.14.4",
    "python-dotenv==1.0.0",
    "python-jose[cryptography]==3.3.0",
    "reportlab==4.0.4",
    "supabase==2.3.4",
]
